feat: 모임 목록 컴포넌트화
